renames & unused cleanup
