+++ crates/assistant/src/prompts.rs
@@ -1,9 +1,3 @@
-pub mod prompt_library2;
+pub mod prompt_library;
 pub mod prompt_manager;
 pub mod prompts;
-pub mod prompts2;
-
-use prompt_library2::*;
-use prompt_manager::*;
-use prompts::*;
-use prompts2::*;
